An R6 class for managing indexed backups

A BackupQueue for indexed backups, e.g. foo.log, foo.1.log, foo.2.log, ...

Details

This class is part of the R6 API of rotor which is intended for developers that want to extend this package. For normal usage, the simpler functional API is recommended (see rotate()).

See also

Super classes

rotor::DirectoryQueue -> rotor::BackupQueue -> BackupQueueIndex

Methods

Public methods

BackupQueueIndex$push()

BackupQueueIndex$prune()

BackupQueueIndex$prune_identical()

BackupQueueIndex$should_rotate()

BackupQueueIndex$pad_index()

BackupQueueIndex$increment_index()

Inherited methods

rotor::DirectoryQueue$set_dir()

rotor::BackupQueue$initialize()

rotor::BackupQueue$print()

rotor::BackupQueue$push_backup()

rotor::BackupQueue$set_backup_dir()

rotor::BackupQueue$set_compression()

rotor::BackupQueue$set_file()

rotor::BackupQueue$set_max_backups()

rotor::BackupQueue$set_origin()

Method push()

Create a new index-stamped backup (e.g. logfile.1.log)

Usage

BackupQueueIndex$push()

Method prune()

Usage

BackupQueueIndex$prune(max_backups = self$max_backups)

Method prune_identical()

Usage

BackupQueueIndex$prune_identical()

Method should_rotate()

Should a file of size be rotated? See size argument of rotate()

Usage

BackupQueueIndex$should_rotate(size, verbose = FALSE)

Returns

TRUE or FALSE

Method pad_index()

Pad the indices in the filenames of indexed backups to the number of digits of the largest index. Usually does not have to be called manually.

Usage

BackupQueueIndex$pad_index()

Method increment_index()

Increment die Indices of all backups by n Usually does not have to be called manually.

Usage

BackupQueueIndex$increment_index(n = 1)

Arguments

n

integer > 0