An R6 class for managing a persistent file-based cache

An R6 class for managing a persistent file-based cache

An R6 class for managing a persistent file-based cache

Details

Cache provides an R6 API for managing a folder that contains cached R objects as .rds files. Cache supports automatic removal of old files if the the cache exceeds a predetermined number of files, total size, or if the files exceed a certain age.

Active bindings

dir

a character scalar. path of the directory in which to store the cache files

max_files

see the compress argument of base::saveRDS(). Note: this differs from the $compress argument of rotate().

max_files

integer scalar: maximum number of files to keep in the cache

max_size

scalar integer, character or Inf. Delete cached files (starting with the oldest) until the total size of the cache is below max_size. Integers are interpreted as bytes. You can pass character vectors that contain a file size suffix like 1k (kilobytes), 3M (megabytes), 4G (gigabytes), 5T (terabytes). Instead of these short forms you can also be explicit and use the IEC suffixes KiB, MiB, GiB, TiB. In Both cases 1 kilobyte is 1024 bytes, 1 megabyte is 1024 kilobytes, etc... .

max_age

a Date scalar: Remove all backups before this date

a character scalar representing a Date in ISO format (e.g. "2019-12-31")

a character scalar representing an Interval in the form "<number> <interval>" (see rotate())

hashfun

NULL or a function to generate a unique hash from the object to be cached (see example). The hash must be a text string that is a valid filename on the target system. If $hashfun is NULL, a storage key must be supplied manually in cache$push().

Methods

Public methods

Cache$new()

Cache$push()

Cache$read()

Cache$remove()

Cache$pop()

Cache$prune()

Cache$purge()

Cache$print()

Cache$set_dir()

Cache$set_max_files()

Cache$set_max_age()

Cache$set_max_size()

Cache$set_compression()

Cache$set_hashfun()

Method new()

Usage

Cache$new(
  dir = dirname(file),
  max_files = Inf,
  max_size = Inf,
  max_age = Inf,
  compression = TRUE,
  hashfun = digest::digest,
  create_dir = TRUE
)

Arguments

create_dir

logical scalar. If TRUE dir is created if it does not exist.

Examples

td &lt;- file.path(tempdir(), "cache-test")

# When using a real hash function as hashfun, identical objects will only
# be added to the cache once
cache_hash &lt;- Cache$new(td, hashfun = digest::digest)
cache_hash$push(iris)
cache_hash$push(iris)
cache_hash$files
cache_hash$purge()

# To override this behaviour use a generator for unique ids, such as uuid
if (requireNamespace("uuid")){
  cache_uid &lt;- Cache$new(td, hashfun = function(x) uuid::UUIDgenerate())
  cache_uid$push(iris)
  cache_uid$push(iris)
  cache_uid$files
  cache_uid$purge()
}

unlink(td, recursive = TRUE)

Method push()

push a new object to the cache

Usage

Cache$push(x, key = self$hashfun(x))

Arguments

key

a character scalar. Key under which to store the cached object. Must be a valid filename. Defaults to beeing generated by $hashfun() but may also be supplied manually.

Returns

a character scalar: the key of the newly added object

Method read()

read a cached file

Usage

Cache$read(key)

Arguments

key

character scalar. key of the cached file to read.

Method remove()

remove a single file from the cache

Usage

Cache$remove(key)

Arguments

key

character scalar. key of the cached file to remove

Method pop()

Read anr remove a single file from the cache

Usage

Cache$pop(key)

Arguments

key

character scalar. key of the cached file to read/remove

Method prune()

Prune the cache

Usage

Cache$prune(
  max_files = self$max_files,
  max_size = self$max_size,
  max_age = self$max_age,
  now = Sys.time()
)

Arguments

max_files, max_size, max_age

see section Active Bindings.

now

a POSIXct datetime scalar. The current time (for max_age)

Method purge()

purge the cache (remove all cached files)

Usage

Cache$purge()

Method print()

Usage

Cache$print()

Method set_dir()

Usage

Cache$set_dir(x, create = TRUE)

Method set_max_files()

Usage

Cache$set_max_files(x)

Method set_max_age()

Usage

Cache$set_max_age(x)

Method set_max_size()

Usage

Cache$set_max_size(x)

Method set_compression()

Usage

Cache$set_compression(x)

Method set_hashfun()

Usage

Cache$set_hashfun(x)

Examples

## ------------------------------------------------
## Method `Cache$new`
## ------------------------------------------------

td <- file.path(tempdir(), "cache-test")

# When using a real hash function as hashfun, identical objects will only
# be added to the cache once
cache_hash <- Cache$new(td, hashfun = digest::digest)

#> creating directory '/tmp/RtmpykgYqg/cache-test'

cache_hash$push(iris)

#> [1] "d3c5d071001b61a9f6131d3004fd0988"

cache_hash$push(iris)

#> [1] "d3c5d071001b61a9f6131d3004fd0988"

cache_hash$files

#>                                                          path
#> 1 /tmp/RtmpykgYqg/cache-test/d3c5d071001b61a9f6131d3004fd0988
#>                                key size isdir mode               mtime
#> 1 d3c5d071001b61a9f6131d3004fd0988 1093 FALSE  664 2020-07-22 11:41:52
#>                 ctime               atime  uid  gid  uname grname
#> 1 2020-07-22 11:41:52 2020-07-22 11:41:52 2000 2000 travis travis

cache_hash$purge()

#> <Cache> 
#> 
#> /tmp/RtmpykgYqg/cache-test	[0 files; 0 B]

# To override this behaviour use a generator for unique ids, such as uuid
if (requireNamespace("uuid")){
  cache_uid <- Cache$new(td, hashfun = function(x) uuid::UUIDgenerate())
  cache_uid$push(iris)
  cache_uid$push(iris)
  cache_uid$files
  cache_uid$purge()
}

#> Loading required namespace: uuid

#> <Cache> 
#> 
#> /tmp/RtmpykgYqg/cache-test	[0 files; 0 B]

unlink(td, recursive = TRUE)