An R6 Class for managing backups (abstract base class)

BackupQueueis an abstract class not intended for direct usage, please refer to BackupQueueIndex, BackupQueueDateTime, BackupQueueDate instead.

Details

This class is part of the R6 API of rotor which is intended for developers that want to extend this package. For normal usage, the simpler functional API is recommended (see rotate()).

See also

Super class

rotor::DirectoryQueue -> BackupQueue

Public fields

dir

character scalar. Directory in which to place the backups.

n

integer scalar. The number of backups that exist for BackupQueue$origin

Active bindings

dir

character scalar. Directory in which to place the backups.

n

integer scalar. The number of backups that exist for BackupQueue$origin

file

character scalar. The file to backup/rotate.

compression

(Optional) compression to use compression argument of rotate().

max_backups

Maximum number/size/age of backups. See max_backups argument of rotate()

has_backups

Returns TRUE if at least one backup of BackupQueue$origin exists All backups of self$origin

Methods

Public methods

BackupQueue$new()

BackupQueue$prune()

BackupQueue$print()

BackupQueue$push_backup()

BackupQueue$set_origin()

BackupQueue$set_compression()

BackupQueue$set_max_backups()

BackupQueue$set_file()

BackupQueue$set_backup_dir()

Inherited methods

rotor::DirectoryQueue$push()

rotor::DirectoryQueue$set_dir()

Method new()

Usage

BackupQueue$new(
  origin,
  dir = dirname(origin),
  max_backups = Inf,
  compression = FALSE,
  backup_dir = NULL
)

Method prune()

Delete all backups except max_backups. See prune_backups().

Usage

BackupQueue$prune(max_backups = self$max_backups)

Method print()

Usage

BackupQueue$print()

Method push_backup()

Usage

BackupQueue$push_backup(...)

Method set_origin()

Usage

BackupQueue$set_origin(x)

Method set_compression()

Usage

BackupQueue$set_compression(x)

Method set_max_backups()

Usage

BackupQueue$set_max_backups(x)

Method set_file()

Usage

BackupQueue$set_file(x)

Method set_backup_dir()

Usage

BackupQueue$set_backup_dir(x)